Update metric values in backup manager (#31)
This PR is second commit of zk backup monitoring implementation. This commit integrates BackupStats with BackupManager.

+++ zookeeper-server/src/main/java/org/apache/zookeeper/server/backup/monitoring/BackupBean.java
@@ -71,7 +71,7 @@ public class BackupBean implements ZKMBeanInfo, BackupMXBean {
   }
 
   @Override
-  public long getNumberOfSnapshotFilesBackedUpLastIteration() {
+  public int getNumberOfSnapshotFilesBackedUpLastIteration() {
     return backupStats.getNumberOfSnapshotFilesBackedUpLastIteration();
   }
